backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
paho-mqtt
pydantic
pytest
httpx
